devop: move read until crlf to utils
