+++ src/libs/stream_handler.rs
@@ -1,4 +1,5 @@
 use crate::libs::redis_cmd::RedisCmd;
+use crate::utils::parser::{ parse_message };
 use std:: {
   io:: { BufRead, BufReader, BufWriter, Result, Write },
   net:: { TcpStream },
@@ -228,76 +229,4 @@ impl<'a> StreamHandler <'a> {
     self.writer.write_all(val.as_bytes()).unwrap();
     self.writer.flush().unwrap();
   }
-}
-
-pub fn parse_message(input: String) -> Result<Message> {
-  match input.chars().next().unwrap() {
-    '+' => parse_simple_string(input),
-    '*' => parse_array(input),
-    '$' => parse_bulk_string(input),
-    _ => panic!("error parse message: no matcher found!")
-  }
-}
-
-pub fn read_until_crlf(input: String) -> Option<String> {
-   let line = input.lines().next();
-   
-   match line {
-    Some(v) => {
-      return Some(v[1..].into());
-    },
-    None => {
-      return None;
-    }
-   }
-}
-
-fn parse_simple_string(input: String) -> Result<Message> {
-  if let Some(v) = read_until_crlf(input) {
-   
-    return Ok(Message::SimpleString(v));
-  } else {
-    return Err(Error::other("oh no!"));
-  }
-}
-
-fn parse_array(input: String) -> Result<Message> {
-  if let Some(v) = read_until_crlf(input.clone()) {
-    let num_str = v.trim();
-
-    let arr_len = num_str.parse::<i64>();
-    // Convert the lines into a vector of owned strings
-    let lines: Vec<String> = input.lines().map(String::from).collect();
-    let mut items: Vec<Message> = vec![];
-
-    let mut inc = 0;
-    for cursor in 0..arr_len.unwrap() {
-             
-      let x = format!("{}\r\n{}", lines[cursor as usize + 1 + inc], lines[cursor as usize + 2 + inc]);
-      let result = parse_message(x);
-          
-      items.push(result.unwrap());
-      inc = inc + 1;
-    }
-    return Ok(Message::Array(items));
-  } 
-
-  return Err(Error::other("oh no!"));
-}
-
-fn parse_bulk_string(input: String) -> Result<Message> {
-  if let Some(v) = read_until_crlf(input.clone()) {
-    let num_str = v.trim();
-
-    let _ = num_str.parse::<i64>();
-
-    for (i, line) in input.lines().enumerate() {
-      if i > 0 {
-        return Ok(Message::BulkString(line.to_string()));
-      }
-
-    }
-  } 
-
-  return Err(Error::other("oh no!"));
 }
